update case 29/6 - cart

+++ case/case.js
@@ -1,25 +1,27 @@
-    function loadWebpage(){
-    localStorage.setItem('iLogIn', 'false');
-}
-
-function information() {
-    document.getElementById('introduce').style.display = 'block'
-}
-
 function home() {
     document.getElementById('introduce').style.display = 'none'
     document.getElementById('display').style.display = 'none'
+    document.getElementById('home').style.display = 'block'
 }
 
+function imfomafion() {
+    document.getElementById('display').style.display = 'none'
+    document.getElementById('home').style.display = 'none'
+    document.getElementById('introduce').style.display = 'block';
+}
 
 function logIn() {
     let checkLogIn = document.getElementById('logIn').innerHTML
     if (checkLogIn === 'Đăng nhập') {
         document.getElementById('formLogIn').style.display = 'block'
     } else {
-        confirm('Bạn có muốn đăng xuất không?')
-        document.getElementById('logIn').innerHTML = 'Đăng nhập'
-        localStorage.setItem('iLogIn', 'false');
+        let comFirm = confirm('Bạn có muốn đăng xuất không?')
+        if (comFirm) {
+            document.getElementById('logIn').innerHTML = 'Đăng nhập'
+            localStorage.setItem('iLogIn', 'false');
+            displayProduct()
+        }
+        document.getElementById('home').style.display = 'block'
     }
 }
 
@@ -33,15 +35,20 @@ function submit() {
             document.getElementById('formLogIn').style.display = 'none'
             localStorage.setItem('iLogIn', 'true')
             document.getElementById('logIn').innerHTML = 'Đăng xuất'
+            reset()
+            displayProduct()
         } else {
             alert('Vui lòng nhập lại mật khẩu!')
         }
     } else alert('Vui lòng nhập lại tên tài khoản!')
+    document.getElementById('home').style.display = 'block'
 }
 
-function cancle(){
-    document.getElementById('formLogIn').style.display = 'none'
+function reset() {
+    document.getElementById('nameLogIn').value = ''
+    document.getElementById('passLogIn').value = ''
 }
+
 class Product {
     name
     price
@@ -83,24 +90,32 @@ let product21 = new Product('Samsung Galaxy Z Fold 4', 40000, 11, 'ảnh/zfold4.
 let arrayProduct = [product1, product2, product3, product4, product5, product6, product7, product8, product9, product10
     , product11, product12, product13, product14, product15, product16, product17, product18, product19, product20, product21]
 
-let iLogIn = localStorage.getItem('iLogIn')
-console.log(iLogIn)
-function displayProduct() {
+
+function displayProduct(type, keyword = '') {
     document.getElementById('introduce').style.display = 'none'
+    document.getElementById('home').style.display = 'none'
     document.getElementById('display').style.display = 'block'
 
+    let iLogIn = localStorage.getItem('iLogIn')
+    let arrayCheckType = arrayProduct.filter(function (product, index, array) {
+        if (type === null) {
+            return product.name.toLowerCase().includes(keyword.toLowerCase())
+        }
+        return product.type === type && product.name.toLowerCase().includes(keyword.toLowerCase())
+    })
+
     let data = ''
-    for (let i = 0; i < arrayProduct.length; i++) {
-        data += '<div class="item" style="float: left">';
-        data += `<img class="img" src=${arrayProduct[i].img} >`
-            + '<h3>' + arrayProduct[i].name + '</h3>'
-            + ' <span>Giá: <strong style="color: red">' + arrayProduct[i].price + '</strong></span><br><br>'
-            + '<span>Số lượng: ' + arrayProduct[i].quantity + '</span><br><br>'
+    for (let i = 0; i < arrayCheckType.length; i++) {
+        data += '<div class="item">';
+        data += `<img class="img" src=${arrayCheckType[i].img} >`
+            + '<h3>' + arrayCheckType[i].name + '</h3>'
+            + ' <span>Giá: <strong style="color: red">' + arrayCheckType[i].price + '</strong></span><br><br>'
+            + '<span>Số lượng: ' + arrayCheckType[i].quantity + '</span><br><br>'
         if (iLogIn === 'true') {
-            data += '<button onclick="editProduct(' + i + ')">Sửa</button>' +
-            ' <button onclick="deleteProduct(' + i + ')">Xóa</button>'
+            data += '<button onclick="editProduct(' + i + ')"><label for="nameEdit">Sửa</label></button>' +
+                ' <button onclick="deleteProduct(' + i + ')">Xóa</button>'
         } else {
-            data += ' <button >Thêm vào giỏ hàng</button><br><br>'
+            data += ' <button onclick="addCart(' + i + ')">Thêm vào giỏ hàng</button><br><br>'
         }
         data += '</div>'
     }
@@ -119,22 +134,29 @@ function addSuccess() {
     let imgAdd = document.getElementById('imgAdd').value
     let typeAdd = document.getElementById('typeAdd').value
 
-    let addProduct = new Product(nameAdd, priceAdd, quantityAdd, imgAdd, typeAdd)
-    console.log(addProduct)
-    if (addProduct.name === '') {
+    let productAdd = new Product(nameAdd, priceAdd, quantityAdd, imgAdd, typeAdd)
+
+    if (productAdd.name === '') {
         alert('Vui lòng thêm tên sản phẩm!')
     } else {
-        arrayProduct.push(addProduct)
+        arrayProduct.push(productAdd)
         alert('Đã thêm thành công!')
         document.getElementById('addProduct').style.display = 'none'
-        displayProduct()
+        displayProduct(null, '')
     }
+    document.getElementById('nameAdd').value = ''
+    document.getElementById('priceAdd').value = ''
+    document.getElementById('quantityAdd').value = ''
+    document.getElementById('imgAdd').value = ''
+    document.getElementById('typeAdd').value = ''
 }
 
 function deleteProduct(id) {
-    confirm('Bạn có chắc chắn muốn xóa hay không?')
-    arrayProduct.splice(id, 1)
-    displayProduct()
+    let confirmDelete = confirm('Bạn có chắc chắn muốn xóa hay không?')
+    if (confirmDelete) {
+        arrayProduct.splice(id, 1)
+        displayProduct(null, '')
+    }
 }
 
 let index;
@@ -155,18 +177,72 @@ function updateProduct() {
     let quantityUpdate = document.getElementById('quantityEdit').value
     let imgUpdate = document.getElementById('imgEdit').value
     let typeUpdate = document.getElementById('typeEdit').value
+    let productUpdate = new Product(nameUpdate, priceUpdate, quantityUpdate, imgUpdate, typeUpdate)
     if (nameUpdate === '') {
         alert('Vui lòng thêm tên sản phẩm!')
     } else {
-        arrayProduct[index].name = nameUpdate
-        arrayProduct[index].price = priceUpdate
-        arrayProduct[index].quantity = quantityUpdate
-        arrayProduct[index].img = imgUpdate
-        arrayProduct[index].type = typeUpdate
+        arrayProduct[index] = productUpdate
         alert('Đã cập nhật sản phẩm thành công!')
     }
     document.getElementById('editProduct').style.display = 'none'
-    displayProduct()
+    displayProduct(null, '')
+}
+
+let arrayCart = []
+
+function addCart(id) {
+    let nameAddCart = arrayProduct[id].name
+    let priceAddCart = arrayProduct[id].price
+    let quantityAddCart = (arrayProduct[id].quantity - (arrayProduct[id].quantity - 1))
+    let imgAddCart = arrayProduct[id].img
+    let typeAddCart = arrayProduct[id].type
+    let productCart = new Product(nameAddCart, priceAddCart, quantityAddCart, imgAddCart, typeAddCart)
+
+
+    let flag = false
+    let count;
+    if (arrayCart.length === 0) {
+        arrayCart.push(productCart)
+        console.log(arrayCart)
+        return arrayCart
+    } else {
+        for (let i = 0; i < arrayCart.length; i++) {
+            if (productCart.name === arrayCart[i].name) {
+                flag = true
+                count = i
+                break;
+            }
+        }
+        if (flag) {
+            arrayCart[count].quantity += 1
+        } else {
+            arrayCart.push(productCart)
+        }
+    }
+    console.log(arrayCart)
+}
+
+
+function displayCart() {
+    let dataCart = '<table>'
+    dataCart += '<tr>' +
+        '<td><b>' + 'Tên sản phẩm' + '</b></td>' +
+        '<td><b>' + ' Giá' + '</b></td>' +
+        '<td><b>' + ' Số lượng' + '</b></td>' +
+        '<td><b>' + 'Thành tiền' + '</b></td>' +
+        '</tr>'
+    for (let i = 0; i < arrayCart.length; i++) {
+        dataCart += '<tr>' +
+            '<td>' + arrayCart[i].name + '</td>' +
+            '<td>' + arrayCart[i].price + '</td>' +
+            '<td>' + '<input type="number" value="' + arrayCart[i].quantity + '" min="0">' + '</td>' +
+            '<td>' + arrayCart[i].price * arrayCart[i].quantity + '</td>' + '<td><button>Xóa</button></td>'
+        dataCart += '</tr>'
+    }
+    dataCart += '<tr style="color: red">' + '<td colspan="3"><b>Tổng thanh toán</b></td>' + '<td><b>' + +'</b></td>' + '</td>'
+    dataCart += '</table>'
+    document.getElementById('displayCart_content').innerHTML = dataCart
+    document.getElementById('displayCart').style.display = 'block'
 }
